Ensure fkey is dropped before removing instance_id
This fixes Bug #904888

On MySQL databases that use InnoDB by default (ie, at the point when
database is being first migrated), extra care needs to be taken to
ensure FKs are dropped before columns because they are not automatically.

Patch from Adam Gandelman and Scott Moser.

+++ nova/db/sqlalchemy/migrate_repo/versions/064_change_instance_id_to_uuid_in_instance_actions.py
@@ -48,6 +48,13 @@ def upgrade(migrate_engine):
         uuid_column.drop()
         raise
 
+    if migrate_engine.name == "mysql":
+        try:
+            migrate_engine.execute("ALTER TABLE instance_actions " \
+                "DROP FOREIGN KEY instance_actions_ibfk_1;")
+        except Exception:  # Don't care, just fail silently.
+            pass
+
     instance_actions.c.instance_id.drop()
